long int en telefono

+++ SampietroFranco_PP/src/duenio.h
@@ -14,7 +14,7 @@ typedef struct
 {
 	int id;
 	char nombre[TAMNOMBREDUENIO];
-	int telefonoContacto;
+	long int telefonoContacto;
 	int estado; //0-Vacio, 1-Ocupado
 }eDuenio;
 
@@ -22,7 +22,7 @@ void duenio_inicializarArray(eDuenio *_arrDuenio, int _tamDuenio);
 int duenio_obtenerIdDesdeLista(eDuenio *_arrDuenio, int _tamDuenio);
 int duenio_buscarDesdeId(eDuenio *_arrDuenio, int _tamDuenio, int _id);
 int duenio_obtenerNombre(eDuenio *_arrDuenio, int _tamDuenio, int _idDuenio, char *_nombreDuenio);
-int duenio_obtenerTelefono(eDuenio *_arrDuenio, int _tamDuenio, int _idDuenio, int *_telDuenio);
+int duenio_obtenerTelefono(eDuenio *_arrDuenio, int _tamDuenio, int _idDuenio, long int *_telDuenio);
 void duenio_mostrarUno(eDuenio _duenio);
 void duenio_mostrarTodos(eDuenio *_arrDuenio, int _tamDuenio);
